Create separate files for each collection
