+++ utils/db/recipes.ts
@@ -1,29 +1,10 @@
 import mongoose from "mongoose";
 
-import Recipe from "../types/client/Recipe";
-import RecipeModel from "../models/RecipeModel";
-import RecipeFilter from "../types/client/RecipeFilter";
-import { isEmptyObject } from "./object";
-
-// MongoDB indexes
-export const Indexes = {
-  RecipeName: "recipe-name",
-} as const;
-
-export const MAX_DOCS = 100;
-
-/**
- * Connect to MongoDB using mongoose
- */
-export const connectToMongoDB = async () => {
-  try {
-    await mongoose.connect(`${process.env.MONGO_URI}`);
-    console.log("Connected to MongoDB");
-  } catch (error) {
-    console.error("Failed to connect to MongoDB:", error);
-    process.exit(1); // exit with failure
-  }
-};
+import { MAX_DOCS, Indexes } from ".";
+import RecipeModel from "../../models/RecipeModel";
+import Recipe from "../../types/client/Recipe";
+import RecipeFilter from "../../types/client/RecipeFilter";
+import { isEmptyObject } from "../object";
 
 /**
  * Write a recipe to MongoDB
